requirements.txt
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
joblib==1.4.2
jsonpickle==3.2.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
networks==0.3.7
networkx==3.3
nltk==3.8.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
progressbar==2.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
regex==2024.5.15
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
